+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - xfer: fix send of data on the DCC chat buffer after `/upgrade` if the buffer was opened before the upgrade ([#2092](https://github.com/weechat/weechat/issues/2092))
 - php: fix return value of function hdata_longlong
 - tcl: fix return value of function hdata_longlong ([#2119](https://github.com/weechat/weechat/issues/2119))
+- relay: fix websocket permessage-deflate extension when the client doesn't send the max window bits parameters
 - relay: fix allocation and reinit of field "client_context_takeover" in websocket deflate structure
 - core: fix detection of libgcrypt ≥ 1.11 ([debian #1071960](https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=1071960))
 - core, relay: fix include directory of libcjson and libzstd